Created the initial structure for the Product List View
-Implemented the initial ProductAdapter class
-Implemented the ProductListFragment
-Implemented the ProductViewModel class with initial mock data
-Adjusted the product_list layout
-Adjusted the product_list_row layout
-Created packages to organize components
-Adjusted reviewRating variable type

+++ app/src/main/java/com/amr/gharseldin/productbrowser/model/data.kt
@@ -29,7 +29,7 @@ data class Product(
     @SerializedName("productImage")
     val productImage: String?,
     @SerializedName("reviewRating")
-    val reviewRating: Int?,
+    val reviewRating: Float?,
     @SerializedName("reviewCount")
     val reviewCount: Int?,
     @SerializedName("inStock")
